apas-agents/requirements.txt
# APAS Agent Layer dependencies
langgraph>=0.2.0
langchain-core>=0.3.0
langchain-openai>=0.2.0

# Shared with ingestion pipeline
chromadb>=0.4.0
sentence-transformers>=2.2.0
python-dotenv>=1.0.0
rich>=13.0.0
